pyproject.toml
[project]
name = "langchain-demo"
version = "0.1.0"
description = "AI apps using LangChain"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.20",
    "langchain-community>=0.0.38",
    "langchain-openai>=0.1.7",
    "python-dotenv>=1.0.1",
]
